requirements.txt
bitcoinrpc
requests
pandas
matplotlib
plotly
pyyaml
